+++ core-ng/src/main/java/core/framework/internal/db/SelectQuery.java
@@ -49,10 +49,12 @@ final class SelectQuery<T> {
         return builder.toString();
     }
 
-    String projectionSQL(String projection, StringBuilder where, String groupBy) {
+    String projectionSQL(String projection, StringBuilder where, String groupBy, String sort, Integer skip, Integer limit) {
         StringBuilder builder = new StringBuilder("SELECT ").append(projection).append(" FROM ").append(table);
         if (where.length() > 0) builder.append(" WHERE ").append(where);
         if (groupBy != null) builder.append(" GROUP BY ").append(groupBy);
+        if (sort != null) builder.append(" ORDER BY ").append(sort);
+        if (skip != null || limit != null) builder.append(" LIMIT ?,?");
         return builder.toString();
     }